+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - **GitHub API** Initial version of the GitHub REST API implementation
 - **Package Handler**, **Test** Extend regex of get_closest_package_tag and tests to handle additional cases
 - **Package Handler** Move logic of nvchecker.toml URL extraction into a separate function and handle additional cases
+- **Test** Add tests for extracting the URL out of the content of nvchecker.toml
 
 ### Bug fixes